7 renaming Cyberkit4SME groups and paths to state_report_management
