+++ app/ssm/common_v2/risk_calculation.py
@@ -40,8 +40,8 @@ from ssm_api_client.models.control_set import ControlSet
 from ssm_api_client.models.threat_dto import ThreatDTO
 
 from app.ssm.ssm_client import SSMClient
-from app.ssm.cyberkit4sme.shortest_path import ShortestPathDataset, ThreatTree
-from app.ssm.cyberkit4sme.shortest_path_graph import plot_rec_csg
+from app.ssm.state_report_management.shortest_path import ShortestPathDataset, ThreatTree
+from app.ssm.state_report_management.shortest_path_graph import plot_rec_csg
 
 #import logging
 #logger = logging.getLogger(__name__)
