+++ app/ssm/state_report_management/bg_shortest_path.py
@@ -51,11 +51,11 @@ from ssm_api_client.models.control_strategy import ControlStrategy
 from ssm_api_client.models.threat import Threat
 from ssm_api_client.models.threat_dto import ThreatDTO
 
-from app.ssm.cyberkit4sme.bg_calculate_risk import calculate_risk
-from app.ssm.cyberkit4sme.bg_process_state_reports import process_state_reports
-from app.ssm.cyberkit4sme.shortest_path import ShortestPathDataset, ThreatTree
-from app.ssm.cyberkit4sme.shortest_path_graph import plot_rec_csg
-from app.ssm.cyberkit4sme.shortest_path_mitigation import ShortestPathMitigation
+from app.ssm.state_report_management.bg_calculate_risk import calculate_risk
+from app.ssm.state_report_management.bg_process_state_reports import process_state_reports
+from app.ssm.state_report_management.shortest_path import ShortestPathDataset, ThreatTree
+from app.ssm.state_report_management.shortest_path_graph import plot_rec_csg
+from app.ssm.state_report_management.shortest_path_mitigation import ShortestPathMitigation
 
 from app.ssm.protego.bg_rollback_utils import restore_twas, clear_twas
