+++ app/ssm/state_report_management/bg_vulnerability_base_twas.py
