+++ app/ssm/state_report_management/bg_vulnerability_openvas.py
@@ -29,7 +29,7 @@ from fastapi.logger import logger
 
 from app.ssm.ssm_client import TWALevel
 
-from app.ssm.cyberkit4sme.bg_vulnerability_base_twas import getVulnerabilityBaseTWAs
+from app.ssm.state_report_management.bg_vulnerability_base_twas import getVulnerabilityBaseTWAs
 
 from app.core.config import NIST_API_KEY
 from time import sleep
