+++ app/ssm/state_report_management/bg_vulnerability_report.py
@@ -40,9 +40,9 @@ from app.crud.store_state_report import store_state_report
 
 from app.ssm.protego.bg_rollback_utils import store_twas
 
-from app.ssm.cyberkit4sme.bg_vulnerability_openvas_parser import OpenVASParser
-from app.ssm.cyberkit4sme.bg_vulnerability_openvas import cvss_base_v2
-from app.ssm.cyberkit4sme.bg_vulnerability_base_twas import getVulnerabilityBaseTWAs
+from app.ssm.state_report_management.bg_vulnerability_openvas_parser import OpenVASParser
+from app.ssm.state_report_management.bg_vulnerability_openvas import cvss_base_v2
+from app.ssm.state_report_management.bg_vulnerability_base_twas import getVulnerabilityBaseTWAs
 
 
 def compose_causation(alert, vulnerability_metrics):
