+++ app/ssm/state_report_management/shortest_path_mitigation.py
@@ -42,8 +42,8 @@ from ssm_api_client.models.control_set import ControlSet
 from ssm_api_client.models.threat_dto import ThreatDTO
 
 from app.ssm.ssm_client import SSMClient
-from app.ssm.cyberkit4sme.shortest_path import ShortestPathDataset, ThreatTree
-from app.ssm.cyberkit4sme.shortest_path_graph import plot_rec_csg
+from app.ssm.state_report_management.shortest_path import ShortestPathDataset, ThreatTree
+from app.ssm.state_report_management.shortest_path_graph import plot_rec_csg
 
 from app.models.ssm.ssm_model import SSMModel
 from app.models.ssm.ssm_model import SSMAsset
